refactor(memory): 下移 undo 语义到插件层
