+++ plugins/default_memory/engine.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import asyncio
 import hashlib
+import json
 import logging
 import time
 from collections.abc import Awaitable, Callable
@@ -61,6 +62,126 @@ def _build_entry_source_ref(base_source_ref: str, entry: str) -> str:
     text = (entry or "").strip()
     digest = hashlib.sha1(text.encode("utf-8")).hexdigest()[:12] if text else "empty"
     return f"{base_source_ref}#h:{digest}"
+
+
+def _source_ref_message_ids(source_ref: str) -> list[str]:
+    raw = str(source_ref or "").strip()
+    if not raw:
+        return []
+    base = raw.split("#", 1)[0].strip()
+    if not base.startswith("["):
+        return []
+    try:
+        loaded = json.loads(base)
+    except json.JSONDecodeError:
+        return []
+    if not isinstance(loaded, list):
+        return []
+    return [str(item).strip() for item in loaded if str(item).strip()]
+
+
+def _undo_store_by_message_sources(
+    store: MemoryStore2,
+    message_ids: list[str],
+    *,
+    dry_run: bool = False,
+) -> dict[str, object]:
+    clean_ids = [str(item).strip() for item in message_ids if str(item).strip()]
+    if not clean_ids:
+        return {"affected_ids": [], "restored_ids": [], "rollback_source_ids": []}
+    target_ids = set(clean_ids)
+    with store._lock:
+        rows = store._db.execute(
+            """
+            SELECT id, source_ref
+            FROM memory_items
+            WHERE COALESCE(source_ref, '') != ''
+            """
+        ).fetchall()
+        affected_ids: set[str] = set()
+        rollback_source_ids: set[str] = set()
+        for item_id, source_ref in rows:
+            source = str(source_ref or "").strip()
+            base_ids = _source_ref_message_ids(source)
+            if source in target_ids:
+                affected_ids.add(str(item_id))
+                rollback_source_ids.add(source)
+                continue
+            if base_ids and target_ids.intersection(base_ids):
+                affected_ids.add(str(item_id))
+                rollback_source_ids.update(base_ids)
+
+        if affected_ids and not dry_run:
+            now = datetime.now().astimezone().isoformat()
+            store._db.executemany(
+                "UPDATE memory_items SET status='superseded', updated_at=? WHERE id=?",
+                [(now, item_id) for item_id in sorted(affected_ids)],
+            )
+        restored_ids = _restore_replacements_for_undo(
+            store,
+            affected_ids,
+            dry_run=dry_run,
+        )
+        if not dry_run:
+            store._db.commit()
+    return {
+        "affected_ids": sorted(affected_ids),
+        "restored_ids": sorted(restored_ids),
+        "rollback_source_ids": sorted(rollback_source_ids),
+    }
+
+
+def _restore_replacements_for_undo(
+    store: MemoryStore2,
+    affected_ids: set[str],
+    *,
+    dry_run: bool = False,
+) -> set[str]:
+    if not affected_ids:
+        return set()
+    sorted_affected = sorted(affected_ids)
+    placeholders = ",".join("?" for _ in sorted_affected)
+    rows = store._db.execute(
+        f"""
+        SELECT DISTINCT old_item_id
+        FROM memory_replacements
+        WHERE new_item_id IN ({placeholders})
+        """,
+        tuple(sorted_affected),
+    ).fetchall()
+    old_ids = {str(row[0]) for row in rows if str(row[0]).strip()}
+    restored: set[str] = set()
+    now = datetime.now().astimezone().isoformat()
+    for old_id in sorted(old_ids):
+        active_replacement = store._db.execute(
+            """
+            SELECT 1
+            FROM memory_replacements r
+            JOIN memory_items m ON m.id = r.new_item_id
+            WHERE r.old_item_id = ?
+              AND r.new_item_id NOT IN ({})
+              AND m.status = 'active'
+            LIMIT 1
+            """.format(placeholders),
+            tuple([old_id, *sorted_affected]),
+        ).fetchone()
+        if active_replacement is not None:
+            continue
+        if dry_run:
+            old_row = store._db.execute(
+                "SELECT 1 FROM memory_items WHERE id=? AND status='superseded'",
+                (old_id,),
+            ).fetchone()
+            if old_row is not None:
+                restored.add(old_id)
+            continue
+        cur = store._db.execute(
+            "UPDATE memory_items SET status='active', updated_at=? WHERE id=? AND status='superseded'",
+            (now, old_id),
+        )
+        if cur.rowcount:
+            restored.add(old_id)
+    return restored
 
 
 def _coerce_emotional_weight(value: object) -> int:
@@ -775,7 +896,8 @@ class DefaultMemoryEngine:
         *,
         dry_run: bool = False,
     ) -> dict[str, object]:
-        return self._require_v2_store().undo_by_message_sources(
+        return _undo_store_by_message_sources(
+            self._require_v2_store(),
             message_ids,
             dry_run=dry_run,
         )
