+++ plugins/plugin_undo/plugin.py
@@ -1,15 +1,28 @@
 from __future__ import annotations
 
 import logging
+from dataclasses import dataclass
+from datetime import datetime
 from typing import Any, cast
 
 from agent.lifecycle.types import BeforeTurnCtx
 from agent.plugins import Plugin
+from agent.prompting import is_context_frame
 
 logger = logging.getLogger("plugin.undo")
 
 _SESSION_SLOT = "session:session"
 _CTX_SLOT = "session:ctx"
+
+
+@dataclass
+class _UndoSessionResult:
+    deleted_ids: list[str]
+    target_user_id: str
+    target_assistant_id: str
+    rollback_index: int
+    last_consolidated_before: int
+    last_consolidated_after: int
 
 
 class UndoCommandModule:
@@ -60,17 +73,31 @@ class PluginUndo(Plugin):
             )
             return _string_list(memory_result.get("rollback_source_ids"))
 
-        result = await session_manager.undo_last_turn(
+        result = await _undo_last_turn(
+            session_manager,
             session_key,
             rollback_source_resolver=resolve_sources,
         )
         if result is None:
             return "没有可撤销的上一轮对话。"
-        memory_result = _undo_memory_sources(
-            getattr(self.context, "memory_engine", None),
-            message_ids_for_memory or result.deleted_ids,
-            dry_run=False,
-        )
+        try:
+            memory_result = _undo_memory_sources(
+                getattr(self.context, "memory_engine", None),
+                message_ids_for_memory or result.deleted_ids,
+                dry_run=False,
+            )
+        except Exception:
+            logger.exception(
+                "undo memory cleanup failed after session delete: session=%s deleted_ids=%s dry_run=%s",
+                session_key,
+                result.deleted_ids,
+                memory_result,
+            )
+            return (
+                "已撤销上一轮对话，但记忆清理失败。"
+                f"\n删除消息：{len(result.deleted_ids)} 条"
+                "\n请查看日志后手动清理对应记忆。"
+            )
         logger.info(
             "undo session=%s deleted=%d memory_superseded=%d memory_restored=%d last=%d->%d",
             session_key,
@@ -86,6 +113,128 @@ class PluginUndo(Plugin):
             f"\n失效记忆：{len(_string_list(memory_result.get('affected_ids')))} 条"
             f"\n恢复旧记忆：{len(_string_list(memory_result.get('restored_ids')))} 条"
         )
+
+
+async def _undo_last_turn(
+    session_manager: Any,
+    session_key: str,
+    *,
+    rollback_source_ids: list[str] | None = None,
+    expected_message_ids: list[str] | None = None,
+    rollback_source_resolver: Any = None,
+) -> _UndoSessionResult | None:
+    async with session_manager._lock(session_key):
+        session = session_manager.get_or_create(session_key)
+        target = _find_last_passive_turn(session.messages)
+        if target is None:
+            return None
+        delete_indices, user_index, assistant_index = target
+        deleted_ids = [
+            str(session.messages[i].get("id") or "")
+            for i in delete_indices
+            if str(session.messages[i].get("id") or "").strip()
+        ]
+        if len(deleted_ids) != len(delete_indices):
+            return None
+        expected = [
+            str(message_id).strip()
+            for message_id in (expected_message_ids or [])
+            if str(message_id).strip()
+        ]
+        if expected and expected != deleted_ids:
+            return None
+        if rollback_source_resolver is not None:
+            rollback_source_ids = rollback_source_resolver(list(deleted_ids))
+        target_user_id = str(session.messages[user_index].get("id") or "")
+        target_assistant_id = str(session.messages[assistant_index].get("id") or "")
+        old_last = max(0, int(session.last_consolidated))
+        rollback_index = _compute_rollback_index(
+            session.messages,
+            delete_indices=delete_indices,
+            old_last_consolidated=old_last,
+            rollback_source_ids=rollback_source_ids or [],
+        )
+        delete_set = set(delete_indices)
+        remaining = [
+            msg for i, msg in enumerate(session.messages) if i not in delete_set
+        ]
+        deleted_before = sum(1 for i in delete_indices if i < rollback_index)
+        new_last = max(0, rollback_index - deleted_before)
+        new_last = min(new_last, len(remaining))
+        deleted_count = session_manager._store.delete_session_messages_and_update_cursor(
+            session.key,
+            ids=deleted_ids,
+            last_consolidated=new_last,
+        )
+        if deleted_count != len(deleted_ids):
+            session_manager.invalidate(session.key)
+            return None
+        session.messages = remaining
+        session.last_consolidated = new_last
+        session.updated_at = datetime.now()
+        session_manager._cache[session.key] = session
+        return _UndoSessionResult(
+            deleted_ids=deleted_ids,
+            target_user_id=target_user_id,
+            target_assistant_id=target_assistant_id,
+            rollback_index=rollback_index,
+            last_consolidated_before=old_last,
+            last_consolidated_after=new_last,
+        )
+
+
+def _is_context_frame_message(message: dict[str, Any]) -> bool:
+    if message.get("role") != "user":
+        return False
+    return is_context_frame(str(message.get("content") or ""))
+
+
+def _is_real_user_message(message: dict[str, Any]) -> bool:
+    return message.get("role") == "user" and not _is_context_frame_message(message)
+
+
+def _is_passive_assistant_message(message: dict[str, Any]) -> bool:
+    return message.get("role") == "assistant" and not bool(message.get("proactive"))
+
+
+def _find_last_passive_turn(
+    messages: list[dict[str, Any]],
+) -> tuple[list[int], int, int] | None:
+    for assistant_index in range(len(messages) - 1, -1, -1):
+        if not _is_passive_assistant_message(messages[assistant_index]):
+            continue
+        user_index = assistant_index - 1
+        while user_index >= 0 and _is_context_frame_message(messages[user_index]):
+            user_index -= 1
+        if user_index < 0 or not _is_real_user_message(messages[user_index]):
+            continue
+        delete_indices = [user_index, assistant_index]
+        context_index = user_index - 1
+        while context_index >= 0 and _is_context_frame_message(messages[context_index]):
+            delete_indices.insert(0, context_index)
+            context_index -= 1
+        return delete_indices, user_index, assistant_index
+    return None
+
+
+def _compute_rollback_index(
+    messages: list[dict[str, Any]],
+    *,
+    delete_indices: list[int],
+    old_last_consolidated: int,
+    rollback_source_ids: list[str],
+) -> int:
+    if not delete_indices:
+        return min(old_last_consolidated, len(messages))
+    rollback_index = min(delete_indices)
+    if rollback_index >= old_last_consolidated:
+        return min(old_last_consolidated, len(messages) - len(delete_indices))
+    source_ids = {str(item).strip() for item in rollback_source_ids if str(item).strip()}
+    for index, message in enumerate(messages):
+        msg_id = str(message.get("id") or "").strip()
+        if msg_id and msg_id in source_ids:
+            rollback_index = min(rollback_index, index)
+    return max(0, min(rollback_index, old_last_consolidated))
 
 
 def _undo_memory_sources(
